requirements.txt
django~=3.2.13
mysqlclient
stripe
python-decouple
boto3
requests
pillow
uWSGI
